# swagger.yaml
openapi: 3.0.3

info:
  title: Kanban Task API
  version: 1.0.0
  description: >
    API для работы с задачами, комментариями и вложениями.
    Используется внешними сервисами и мобильными приложениями.

servers:
  - url: https://your-domain.com/api
    description: Production API
  - url: http://localhost/api
    description: Local development

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token

security:
  - BearerAuth: []



tags:
  - name: Tasks (External API)
    description: Методы для создания задач через внешний API
  - name: Task Comments
    description: Работа с комментариями задач
  - name: Task Attachments
    description: Работа с вложениями задач

paths:

  /task/create:
    post:
      tags:
        - Tasks (External API)
      summary: Создание задачи через внешний API
      description: Универсальный метод создания задачи. Принимает данные задачи, тип, UUID доски и номер колонки.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                board_uuid:
                  type: string
                  example: "123e4567-e89b-12d3-a456-426614174000"
                thread:
                  type: integer
                  example: 2
                type:
                  type: string
                  example: "order"
                payload:
                  type: object
                  example:
                    title: "Новая задача"
                    description: "Описание задачи"
      responses:
        '200':
          description: Задача успешно создана
        '400':
          description: Ошибка валидации

  /task/{task}/comments:
    get:
      tags:
        - Task Comments
      summary: Получить список комментариев задачи
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список комментариев
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    author:
                      type: string
                    text:
                      type: string
                    attachments:
                      type: array
                      items:
                        type: object
                        properties:
                          name:
                            type: string
                          path:
                            type: string
                          mime:
                            type: string
                          size:
                            type: integer
                    created_at:
                      type: string
                      format: date-time

  /task/{task}/comment:
    post:
      tags:
        - Task Comments
      summary: Добавить комментарий к задаче
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                author:
                  type: string
                  example: "Алексей"
                text:
                  type: string
                  example: "Комментарий к задаче"
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Комментарий создан

  /task/{task}/attachments:
    get:
      tags:
        - Task Attachments
      summary: Получить список вложений задачи
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список вложений
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    path:
                      type: string
                    mime:
                      type: string
                    size:
                      type: integer

    post:
      tags:
        - Task Attachments
      summary: Добавить вложения к задаче
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Вложения добавлены